debugs

+++ ty_game_lib/game_stuff/TransRegenEffectStatus.cs
@@ -14,21 +14,21 @@ namespace game_stuff
             ShieldToArmor = 0;
         }
 
-        private int HpToArmor { get; set; }
-        private int HpToShield { get; set; }
-        private int ArmorToShield { get; set; }
-        private int ArmorToHp { get; set; }
-        private int ShieldToHp { get; set; }
-        private int ShieldToArmor { get; set; }
+        private float HpToArmor { get; set; }
+        private float HpToShield { get; set; }
+        private float ArmorToShield { get; set; }
+        private float ArmorToHp { get; set; }
+        private float ShieldToHp { get; set; }
+        private float ShieldToArmor { get; set; }
 
         public void TransRegenerationEffectChange(float[] transRegeneration)
         {
-            HpToArmor = (int) transRegeneration[0];
-            HpToShield = (int) transRegeneration[1];
-            ArmorToShield = (int) transRegeneration[2];
-            ArmorToHp = (int) transRegeneration[3];
-            ShieldToHp = (int) transRegeneration[4];
-            ShieldToArmor = (int) transRegeneration[5];
+            HpToArmor = transRegeneration[0];
+            HpToShield = transRegeneration[1];
+            ArmorToShield = transRegeneration[2];
+            ArmorToHp = transRegeneration[3];
+            ShieldToHp = transRegeneration[4];
+            ShieldToArmor = transRegeneration[5];
 #if DEBUG
             Console.Out.WriteLine(
                 $"now regen attr~ a2s:{ArmorToShield} h2s:{HpToShield} h2a:{HpToArmor} s2a:{HpToArmor} s2h:{ShieldToHp} a2h:{ArmorToHp}");
@@ -38,9 +38,9 @@ namespace game_stuff
         public void GetTransValue(int sLoss, int aLoss, int hLoss, out int sR, out int armorR,
             out int hpR)
         {
-            hpR = ArmorToHp * aLoss + ShieldToHp * sLoss;
-            armorR = ShieldToArmor * sLoss + HpToArmor * hLoss;
-            sR = ArmorToShield * aLoss + HpToShield * hLoss;
+            hpR = (int) (ArmorToHp * aLoss + ShieldToHp * sLoss);
+            armorR = (int) (ShieldToArmor * sLoss + HpToArmor * hLoss);
+            sR = (int) (ArmorToShield * aLoss + HpToShield * hLoss);
 
 #if DEBUG
             Console.Out.WriteLine($"take loss are s:{sLoss} a:{aLoss} h:{hLoss}");
